requirements.txt
numpy~=1.20.3
pandas~=1.3.2
matplotlib~=3.4.2
seaborn~=0.11.2
scikit-learn~=0.24.2
PyYAML~=5.1.2
lightgbm~=3.1.1
optuna~=2.9.1
tqdm~=4.62.1
setuptools~=52.0.0
scipy~=1.6.2
torch~=1.7.1
ipython~=7.26.0
requests~=2.25.1
pyspark~=2.4.5
shap~=0.39.0
ortools~=9.3.10497
aequitas~=0.42.0
joblib~=1.0.1